+++ scripts/install_incus.sh
@@ -1,0 +1,19 @@
+#!/bin/bash
+set -eux
+
+sudo curl -fsSL https://pkgs.zabbly.com/key.asc -o /etc/apt/keyrings/zabbly.asc
+sudo sh -c 'cat <<EOF > /etc/apt/sources.list.d/zabbly-incus-stable.sources
+Enabled: yes
+Types: deb
+URIs: https://pkgs.zabbly.com/incus/stable
+Suites: $(. /etc/os-release && echo ${VERSION_CODENAME})
+Components: main
+Architectures: $(dpkg --print-architecture)
+Signed-By: /etc/apt/keyrings/zabbly.asc
+
+EOF'
+
+sudo apt-get update && sudo apt-get install --yes incus
+
+sudo incus admin init --auto
+sudo usermod -a -G incus $(whoami)
